fix: rename init files, remove hubble files
