+++ jinahub/indexers/searcher/FaissSearcher/config.yml
@@ -3,7 +3,6 @@ with:
   index_key: 'IVF10,PQ4'
 metas:
   py_modules:
-    # - you can put more dependencies here
-    - __init__.py
+    - faiss_searcher.py
   name: faiss_searcher
