+++ jinahub/indexers/searcher/HnswlibSearcher/config.yml
@@ -3,7 +3,6 @@ with:
   distance: 'cosine'
 metas:
   py_modules:
-    # - you can put more dependencies here
-    - __init__.py
+    - hnswlib_searcher.py
   name: hnswlib_searcher
