+++ jinahub/indexers/searcher/HnswlibSearcher/tests/test_hnswlibsearcher.py
@@ -9,7 +9,7 @@ from jina import Document, DocumentArray, Flow
 from jina.executors.metas import get_default_metas
 from jina_commons.indexers.dump import import_vectors
 
-from .. import HnswlibSearcher
+from ..hnswlib_searcher import HnswlibSearcher
 
 # fix the seed here
 np.random.seed(500)
