+++ jinahub/indexers/searcher/NumpySearcher/config.yml
@@ -3,7 +3,6 @@ with:
   {}
 metas:
   py_modules:
-    # - you can put more dependencies here
-    - __init__.py
+    - numpy_searcher.py
   name: numpy_searcher
