--- jinahub/indexers/searcher/compound/NumpyLMDBSearcher/__init__.py
@@ -1,1 +1,0 @@
-from .npfile import NumpyLMDBSearcher
