--- jinahub/indexers/searcher/compound/NumpyPostgresSearcher/__init__.py
@@ -1,1 +1,0 @@
-from .nppsql import NumpyPostgresSearcher
