+++ jinahub/indexers/storage/LMDBStorage/config.yml
@@ -1,7 +1,7 @@
 jtype: LMDBStorage
 metas:
   py_modules:
-    - __init__.py
+    - lmdb_storage.py
   workspace: workspace/
   name: lmdb
