+++ jinahub/indexers/storage/PostgreSQLStorage/config.yml
@@ -3,7 +3,7 @@ with:
   {}
 metas:
   py_modules:
-    - __init__.py
+    - postgres.py
     - postgreshandler.py
   workspace: workspace/
   name: psql
